How to use Python External and built in modules
Here random module demonstration is done which is an inbuilt function. To look into the source code
import random and then press ctrl and then click on random. randon.py will open in pycharm
or else refer python documentation for python modules.

For using modules which are not available inbuilt, they need to be downloaded and installed into the system.
Either use pip install in command prompt or if you're using pycharm go to terminal and write pip install <module name>