Exercise6-Game Development-Snake Water Gun
This is the first ever game I am developing using Python. I know it's a silly game but everything's silly when you
start something.
Snake Water Gun- It's a multiplayer game where
1. If someone says Snake and the other person says water, snake drinks the water and snake wins.
2. If someone says water and the other says gun, the gun drowns in the water and water wins.
3. If someone says gun and the other says snake, the gun kills the snake and gun wins-- Par naagin intekaam legi bhai.

So this game will be between the user and the computer. It will be played 10 times and whoever with most number of wins,
wins the game.